+++ odoo/local-src/specific_partner/__manifest__.py
@@ -16,6 +16,7 @@
         'data/partner.alcyon_category.csv',
         'security/ir.model.access.csv',
         'views/res_partner.xml',
+        'views/legal_entity.xml',
     ],
     'installable': True,
 }