+++ odoo/local-src/specific_partner/models/__init__.py
@@ -4,3 +4,4 @@
 
 from . import partner_alcyon_category
 from . import res_partner
+from . import legal_entity